pyproject.toml
[tool.poetry]
name = "lido-otc-seller"
version = "0.1.0"
description = "TBD"
authors = ["Lido dev team"]
license = "MIT"
readme = "README.md"
homepage = "https://lido.fi"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
eth-brownie = "^1.19.1"
dotmap = "^1.3.30"

[tool.poetry.dev-dependencies]
black = "~22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 160
target-version = ['py310']
include = '\.pyi?$'
